Major level-update
Overhauled the level-system completely: A level is now divided up in separate parts, called acts. Everytime an act is finished, a transition is being played. This whole process is being handled by a state-machine. Everything is implemented to be as customizable as possible, so creating new levels and acts shouldn't be a problem.