Added modular enemy-creation
Added a new way to create enemies using composition instead of inheritance. Created a base enemy (CompositionEnemy) which can be used to test and create modular enemies. Added various functionalities to such enemies, which have been implementend as components. Created a modular version of the FlamingSkull-enemy by using the aforementioned concepts. This change will make the creation of enemies a lot more extensive, customizable and the code more readable